Improved docs removed dead code (and fixed some linting) (#1077)
* Fixed compiler warnings

* Added description for filter_perf

* generated docs

* Fixed more compiler issues

* removed unused folder

* removed old travis build

* removed old build script

* Not used anymore

* removed more unused scripts

* removed unused modules

* Improved filter parameter docs

* whoops

* More compiler warnings

* Whoops

+++ include/collectd/collectd_packet.hpp
@@ -261,7 +261,7 @@ struct collectd_builder {
     if (svalue.first == "derive") {
       std::string vkey = svalue.second;
       if (vkey.size() > 0 && vkey[0] >= '0' && vkey[0] <= '9')
-        metric.derives.push_back(str::stox<double>(svalue.second, 0));
+        metric.derives.push_back(static_cast<long long>(str::stox<double>(svalue.second, 0)));
       else
         metric.derives.push_back(str::stox<unsigned long long>(metrics[svalue.second], 0));
     }
